Add files via upload
If you would like to create an apk for use on Android, please change the main activity into this java file. And don't forget to add an asset file to store all the HTML files. Also, Android Manifest file need to be changed so that all permissions could be obtained.